requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
asyncpg>=0.29.0
telethon==1.33.0
beautifulsoup4==4.12.2
requests==2.31.0
Pillow==10.1.0
apscheduler==3.10.4
python-magic==0.4.27
aiohttp==3.9.1
sqlalchemy==2.0.23
alembic==1.12.1
pydantic-settings==2.1.0
aiosqlite==0.19.0
httpx>=0.25.0
